Left side bar uses <TagsFollowed /> component now.
