--- app/javascript/packs/homePage.js
@@ -1,21 +1,0 @@
-function toggleListingsMinimization() {
-  if (document.body.classList.contains('config_minimize_newest_listings')) {
-    // Un-minimize
-    localStorage.setItem('config_minimize_newest_listings', 'no');
-    document.body.classList.remove('config_minimize_newest_listings');
-  } else {
-    // Minimize
-    localStorage.setItem('config_minimize_newest_listings', 'yes');
-    document.body.classList.add('config_minimize_newest_listings');
-  }
-}
-
-const sidebarListingsMinimizeButton = document.getElementById(
-  'sidebar-listings-widget-minimize-button',
-);
-if (sidebarListingsMinimizeButton) {
-  sidebarListingsMinimizeButton.addEventListener(
-    'click',
-    toggleListingsMinimization,
-  );
-}
