+++ app/views/articles/_sidebar_nav.html.erb
@@ -48,7 +48,7 @@
       <% end %>
     </div>
     <div class="sidebar-nav-block-inner">
-      <div id="sidebar-nav-followed-tags" class="sidebar-nav-followed-tags"></div>
+      <div id="sidebar-nav-followed-tags-ctn"></div>
       <div id="sidebar-nav-default-tags" class="sidebar-nav-default-tags <%= "showing" unless user_signed_in? %>">
         <% if user_signed_in? %>
           <div class="sidebar-nav-subheader tag-separator" id="tag-separator">Other Popular Tags</div>
